refactor: use api name
