+++ src/api/index.ts
@@ -1,27 +1,9 @@
-import config from "@/gtag/config";
-import event from "@/gtag/event";
-import linker from "@/gtag/linker";
-import pageview from "@/gtag/pageview";
-import purchase from "@/gtag/purchase";
-import query from "@/gtag/query";
-import refund from "@/gtag/refund";
-import screenview from "@/gtag/screenview";
-import time from "@/gtag/time";
-
-const api = {
-  config,
-  screenview,
-  pageview,
-  query,
-  time,
-  event,
-  refund,
-  purchase,
-  linker,
-};
-
-export type UseGtagReturn = typeof api;
-
-export function useGtag(): UseGtagReturn {
-  return api;
-}
+export { default as config } from "@/api/config";
+export { default as event } from "@/api/event";
+export { default as linker } from "@/api/linker";
+export { default as pageview } from "@/api/pageview";
+export { default as purchase } from "@/api/purchase";
+export { default as query } from "@/api/query";
+export { default as refund } from "@/api/refund";
+export { default as screenview } from "@/api/screenview";
+export { default as time } from "@/api/time";
