+++ src/api/purchase.ts
@@ -1,4 +1,4 @@
-import event from "@/gtag/event";
+import event from "@/api/event";
 
 export default function purchase(params: Gtag.EventParams) {
   event("purchase", params);
