+++ src/api/refund.ts
@@ -1,4 +1,4 @@
-import event from "@/gtag/event";
+import event from "@/api/event";
 
 export default function refund(params: Gtag.EventParams) {
   event("refund", params);
