+++ src/api/time.test.ts
@@ -1,7 +1,7 @@
-import query from "@/gtag/query";
-import time from "@/gtag/time";
+import query from "@/api/query";
+import time from "@/api/time";
 
-vi.mock("@/gtag/query");
+vi.mock("@/api/query");
 
 describe("time", () => {
   it("should use the timing_complete event", () => {
