+++ src/api/time.ts
@@ -1,4 +1,4 @@
-import event from "@/gtag/event";
+import event from "@/api/event";
 
 export default function time(params: Gtag.EventParams) {
   event("timing_complete", params);
