+++ src/composable/useGtag.test.ts
@@ -1,0 +1,8 @@
+import * as api from "@/api";
+import useGtag from "./useGtag";
+
+describe("gtag", () => {
+  it("should return useGtag composable", () => {
+    expect(useGtag()).toEqual(api);
+  });
+});
