+++ src/composable/useGtag.ts
@@ -1,0 +1,7 @@
+import * as api from "@/api";
+
+export type UseGtagReturn = typeof api;
+
+export default function useGtag(): UseGtagReturn {
+  return api;
+}
